cleanup + comments
not sure what happened to the gitignore; i feel like i can probably lose a lot of it
