+++ .gitignore
@@ -3,8 +3,8 @@
 # Logs
 
 logs
-_.log
-npm-debug.log_
+*.log
+npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 lerna-debug.log*
@@ -15,87 +15,68 @@ lerna-debug.log*
 report.[0-9]_.[0-9]_.[0-9]_.[0-9]_.json
 
 # Runtime data
-
 pids
-_.pid
-_.seed
-\*.pid.lock
+*.pid
+*.seed
+*.pid.lock
 
 # Directory for instrumented libs generated by jscoverage/JSCover
-
 lib-cov
 
 # Coverage directory used by tools like istanbul
-
 coverage
-\*.lcov
+*.lcov
 
 # nyc test coverage
-
 .nyc_output
 
 # Grunt intermediate storage (https://gruntjs.com/creating-plugins#storing-task-files)
-
 .grunt
 
 # Bower dependency directory (https://bower.io/)
-
 bower_components
 
 # node-waf configuration
-
 .lock-wscript
 
 # Compiled binary addons (https://nodejs.org/api/addons.html)
-
 build/Release
 
 # Dependency directories
-
 node_modules/
 jspm_packages/
 
 # Snowpack dependency directory (https://snowpack.dev/)
-
 web_modules/
 
 # TypeScript cache
-
 \*.tsbuildinfo
 
 # Optional npm cache directory
-
 .npm
 
 # Optional eslint cache
-
 .eslintcache
 
 # Optional stylelint cache
-
 .stylelintcache
 
 # Microbundle cache
-
 .rpt2_cache/
 .rts2_cache_cjs/
 .rts2_cache_es/
 .rts2_cache_umd/
 
 # Optional REPL history
-
 .node_repl_history
 
 # Output of 'npm pack'
-
-\*.tgz
+*.tgz
 
 # Yarn Integrity file
-
 .yarn-integrity
 
 # dotenv environment variable files
-
 .env
 .env.development.local
 .env.test.local
@@ -104,72 +85,59 @@ web_modules/
 
 config.ts
 
-# parcel-bundler cache (https://parceljs.org/)
+# Typescript output
+/build
 
+# parcel-bundler cache (https://parceljs.org/)
 .cache
 .parcel-cache
 
 # Next.js build output
-
 .next
 out
 
 # Nuxt.js build / generate output
-
 .nuxt
 dist
 
 # Gatsby files
-
 .cache/
 
 # Comment in the public line in if your project uses Gatsby and not Next.js
-
 # https://nextjs.org/blog/next-9-1#public-directory-support
-
 # public
 
 # vuepress build output
-
 .vuepress/dist
 
 # vuepress v2.x temp and cache directory
-
 .temp
 .cache
 
 # Docusaurus cache and generated files
-
 .docusaurus
 
 # Serverless directories
-
 .serverless/
 
 # FuseBox cache
-
 .fusebox/
 
 # DynamoDB Local files
-
 .dynamodb/
 
 # TernJS port file
-
 .tern-port
 
 # Stores VSCode versions used for testing VSCode extensions
-
 .vscode-test
 
 # VSCode editor settings
-
-.vscode
+# .vscode
 
 # yarn v2
-
 .yarn/cache
 .yarn/unplugged
 .yarn/build-state.yml
 .yarn/install-state.gz
-.pnp.\*
+.pnp.*
