+++ .vscode/settings.json
@@ -2,12 +2,15 @@
     "editor.codeActionsOnSave": {
       "source.fixAll.eslint": "explicit"
     },
+    "editor.defaultFormatter": "vscode.typescript-language-features",
     "editor.detectIndentation": false,
     "editor.insertSpaces": true,
     "editor.tabSize": 2,
-    "files.insertFinalNewline": true,
-    "typescript.preferences.quoteStyle": "single",
+
     "eslint.format.enable": true,
-    "editor.defaultFormatter": "vscode.typescript-language-features"
+
+    "files.insertFinalNewline": true,
+
+    "typescript.preferences.quoteStyle": "single",
   }
