+++ src/lib/bot/index.ts
@@ -1,5 +1,7 @@
 import type { KnownBlock } from '@slack/web-api';
 
+/** @todo who does this belong to? */
+
 /** 
  * Generate daily puzzle blocks 
  * @see https://api.slack.com/interactivity/slash-commands#responding_immediate_response
