+++ src/lib/lichess/index.ts
@@ -7,7 +7,7 @@ const LichessApi = wretch(`https://lichess.org/api`)
  * Responsible for interacting with the Lichess web api
  */
 export default {
-    
+  /** */
   fetchDailyPuzzle: async () => {
     console.info('Fetching daily puzzle data...');
 
@@ -22,6 +22,5 @@ export default {
       puzzleUrl: `https://lichess.org/training/${puzzle.id}`,
       puzzleThumbUrl: `https://lichess1.org/training/export/gif/thumbnail/${puzzle.id}.gif`,
     };
-
   }, 
 };
