+++ src/lib/slack/index.ts
@@ -2,8 +2,7 @@ import wretch from 'wretch';
 import QueryStringAddon from "wretch/addons/queryString"
 
 import config from '../../config';
-import { constructHref, unix } from "../utils";
-import hmac from "../hmac";
+import { constructHref, hmac, unix } from "../utils";
 import {
   parseBotCredentialResponse,
   parseHeaders,
@@ -12,26 +11,6 @@ import {
 
 const SlackApi = wretch('https://slack.com/api')
   .addon(QueryStringAddon)
-
-/**
- * Client
- * @see https://api.slack.com/authentication/oauth-v2#asking
- * @returns 
- */
-const getOAuthRedirectUrl = (uid: string) => {
-  // user tz?
-  const APP_SCOPES = [
-    'commands',
-    'users:read',
-  ];
-
-  return constructHref('https://slack.com/oauth/v2/authorize', {
-    client_id: config.SLACK_CLIENT_ID,
-    scope: APP_SCOPES.join(),
-    state: uid,
-    redirect_uri: config.REGISTRATION_URL,
-  });
-};
 
 /**
  * Protect against replay attacks by enforcing X
@@ -50,61 +29,69 @@ const validateTimestamp = (timestamp: string) => {
 }
 
 /**
- * Validate signature using hmac
- * @see https://api.slack.com/authentication/verifying-requests-from-slack
- */
-const verifyRequest = (request: {
-  headers: Headers;
-  body: ReadableStream<any> | null;
-}) => {
-  const { signature, timestamp } = parseHeaders(request.headers);
-
-  const signatureData = `v0:${timestamp}:${JSON.stringify(request.body)}`;
-  const expectedSignature = hmac.createDigest(config.SLACK_CLIENT_SECRET, signatureData);
-  
-  return {
-    timestampIsValid: validateTimestamp(timestamp),
-    signatureIsValid: hmac.safeCompareDigests(expectedSignature, signature),
-  }
-}
-
-const getTimeZone = async (userId: string) => {
-  return await SlackApi
-    .auth(`Bearer ${config.SLACK_BOT_TOKEN}`)
-    .query({ user: userId })
-    .get('/users.info')
-    .json(parseTimeZone)
-}
-
-const registerBot = async (code: string) => {
-  /**
-   * @see https://api.slack.com/methods/oauth.v2.access
-   */
-  return await SlackApi
-    .auth(`Basic ${config.SLACK_CLIENT_ID}:${config.SLACK_CLIENT_SECRET}`)
-    .query({ 
-      code,
-      redirect_uri: config.REGISTRATION_URL
-    })
-    .post('/oauth.v2.access')
-    .json(parseBotCredentialResponse)
-}
-
-const unregisterBot = async (token: string) => {
-
-  return await SlackApi
-    .query({ token })
-    .post('/auth.revoke')
-    .json()
-}
-
-/**
  * Use to parse and verify requests
  */
 export default {
-  getOAuthRedirectUrl,
-  getTimeZone,
-  registerBot,
-  unregisterBot,
-  verifyRequest,
+  /**
+   * @see https://api.slack.com/authentication/oauth-v2#asking
+   */
+  getOAuthRedirectUrl: (uid: string) => {
+    // user tz?
+    const APP_SCOPES = [
+      'commands',
+      'users:read',
+    ];
+  
+    return constructHref('https://slack.com/oauth/v2/authorize', {
+      client_id: config.SLACK_CLIENT_ID,
+      scope: APP_SCOPES.join(),
+      state: uid,
+      redirect_uri: config.REGISTRATION_URL,
+    });
+  },
+
+  getTimeZone: async (userId: string) => {
+    return await SlackApi
+      .auth(`Bearer ${config.SLACK_BOT_TOKEN}`)
+      .query({ user: userId })
+      .get('/users.info')
+      .json(parseTimeZone)
+  },
+
+  registerBot: async (code: string) => {
+    /** @see https://api.slack.com/methods/oauth.v2.access */
+    return await SlackApi
+      .auth(`Basic ${config.SLACK_CLIENT_ID}:${config.SLACK_CLIENT_SECRET}`)
+      .query({ 
+        code,
+        redirect_uri: config.REGISTRATION_URL
+      })
+      .post('/oauth.v2.access')
+      .json(parseBotCredentialResponse)
+  },
+
+  unregisterBot: async (token: string) => {
+    return await SlackApi
+      .query({ token })
+      .post('/auth.revoke')
+      .json()
+  },
+  /**
+   * Validate signature using hmac
+   * @see https://api.slack.com/authentication/verifying-requests-from-slack
+   */
+  verifyRequest: (request: {
+    headers: Headers;
+    body: ReadableStream<any> | null;
+  }) => {
+    const { signature, timestamp } = parseHeaders(request.headers);
+  
+    const signatureData = `v0:${timestamp}:${JSON.stringify(request.body)}`;
+    const expectedSignature = hmac.createDigest(config.SLACK_CLIENT_SECRET, signatureData);
+    
+    return {
+      timestampIsValid: validateTimestamp(timestamp),
+      signatureIsValid: hmac.safeCompareDigests(expectedSignature, signature),
+    }
+  },
 }
