+++ src/lib/utils.ts
@@ -1,4 +1,4 @@
-
+import { createHmac, timingSafeEqual } from 'crypto';
 
 export const constructHref = (
   baseUrl: string, 
@@ -18,6 +18,40 @@ export const constructHref = (
   return url.href;
 };
 
+/**
+ * A utility exposing the logic required to verify Slack signatures
+ * using HMAC (Hash-based Message Authentication Code)
+ * @todo update links to bun
+ * @see https://nodejs.org/docs/latest-v6.x/api/crypto.html#crypto_class_hmac
+ * @see https://nodejs.org/docs/latest-v6.x/api/crypto.html#crypto_crypto_timingsafeequal_a_b
+ */
+export const hmac = {
+  /**
+   * Create a new hex-encoded HMAC using the provided secret and data
+   * @param secret Secret key
+   * @param data String data to include in token
+   * @returns HMAC hex digest string
+   */
+  createDigest: (secret: string, data: string) => {
+    return createHmac('sha256', secret)
+      .update(data)
+      .digest('hex');
+  },
+  /**
+   * Compare two strings using a time-constant algo, preventing malicious
+   * users from using failure time delta to deduce character mismatch
+   * @param hmacDigestA
+   * @param hmacDigestB 
+   * @returns boolean representing equality
+   */
+  safeCompareDigests: (hmacDigestA: string, hmacDigestB: string) => {
+    const hmacBufferA = Buffer.from(hmacDigestA);
+    const hmacBufferB = Buffer.from(hmacDigestB);
+   
+    return timingSafeEqual(hmacBufferA, hmacBufferB);
+  }
+}
+
 export const unix = {
   fromDate: (date: Date) => {
     return `${Math.floor(date.valueOf() / 1000)}`
@@ -28,15 +62,3 @@ export const unix = {
     return epochSeconds * 1000
   }
 }
-
-
-
-// import { zonedTimeToUtc, utcToZonedTime } from 'date-fns-tz'
-
-// const stripTimezone = () => {
-//   return zonedTimeToUtc(date, timeZone)
-// }
-
-// const setTimezone = () => {
-//   return utcToZonedTime(date, timeZone)
-// }
